altera assinatura das funções para o tipo do tamanho do vetor

+++ include/busca.h
@@ -1,7 +1,7 @@
 #pragma once
 
-int buscaBinaria(int *v, long n, int elem, long long *passos);
+int buscaBinaria(int *v, int n, int elem, long long *passos);
 
-int buscaLinear(int *v, long n, int elem, long long *passos);
+int buscaLinear(int *v, int n, int elem, long long *passos);
 
-int buscaOrdenada(int *v, long n, int elem, long long *passos);
+int buscaOrdenada(int *v, int n, int elem, long long *passos);
